+++ client/normalize-rootlink.js
@@ -1,0 +1,5 @@
+import {default as normalize} from './normalize-2base.js';
+
+export default function normalizerootLink(url) {
+  return normalize(url, '/');
+};